break up index.evals.ts and utils into smaller files
