+++ evals/index.eval.ts
@@ -1,8 +1,8 @@
 /**
  * This script orchestrates the running of evaluations against a set of tasks.
- * It leverages the "braintrust" evaluation framework to run multiple testcases
- * (each testcase representing a given task-model combination) and then aggregates
- * the results, producing a summary of passes, failures, and categorized success rates.
+ * It braintrust to run multiple testcases (each testcase representing a
+ * given task-model combination) and then aggregates the results, producing
+ * a summary of passes, failures, and categorized success rates.
  *
  * Overview:
  * - Reads a configuration file `evals.config.json` to determine what tasks (evaluations)
@@ -12,160 +12,23 @@
  * - Runs each selected task against each selected model in parallel, collecting results.
  * - Saves a summary of the evaluation results to `eval-summary.json`.
  */
-
 import fs from "fs";
 import path from "path";
 import process from "process";
-import { Eval } from "braintrust";
-import {
-  EvalCategorySchema,
-  EvalFunction,
-  SummaryResult,
-  Testcase,
-} from "../types/evals";
-import { AvailableModel, AvailableModelSchema } from "../types/model";
-import { EvalLogger, env, generateExperimentName } from "./utils";
+import { env } from "./env";
+import { generateExperimentName } from "./utils";
 import { exactMatch, errorMatch } from "./scoring";
+import { tasksByName, MODELS } from "./taskConfig";
+import { filterByCategory, filterByEvalName, useTextExtract } from "./args";
+import { Eval } from "braintrust";
+import { EvalFunction, SummaryResult, Testcase } from "../types/evals";
+import { EvalLogger } from "./logger";
+import { AvailableModel } from "../types/model";
+import dotenv from "dotenv";
+dotenv.config();
 
 const MAX_CONCURRENCY = 20;
 const TRIAL_COUNT = 5;
-
-// Extract command-line arguments passed to this script.
-const args = process.argv.slice(2);
-
-// The configuration file `evals.config.json` contains a list of tasks and their associated categories.
-const configPath = path.join(__dirname, "evals.config.json");
-const config = JSON.parse(fs.readFileSync(configPath, "utf-8"));
-
-/**
- * The default categories of evaluations to run if none is specified.
- * These categories represent different styles or types of tasks.
- */
-const DEFAULT_EVAL_CATEGORIES = process.env.EVAL_CATEGORIES
-  ? process.env.EVAL_CATEGORIES.split(",")
-  : [
-      "observe",
-      "act",
-      "combination",
-      "extract",
-      "experimental",
-      "text_extract",
-    ];
-
-/**
- * Determine which extraction method to use for tasks that involve extraction.
- * By default, "domExtract" is used. However, if a `--extract-method=<method>`
- * argument is provided, it will override the default.
- */
-let extractMethod = "domExtract";
-const extractMethodArg = args.find((arg) =>
-  arg.startsWith("--extract-method="),
-);
-if (extractMethodArg) {
-  extractMethod = extractMethodArg.split("=")[1];
-}
-
-// Set the extraction method in the process environment so tasks can reference it.
-process.env.EXTRACT_METHOD = extractMethod;
-const useTextExtract = process.env.EXTRACT_METHOD === "textExtract";
-
-/**
- * Variables for filtering which tasks to run:
- * - `filterByCategory`: if provided, only tasks that belong to this category will be run.
- * - `filterByEvalName`: if provided, only the task with this name will be run.
- */
-let filterByCategory: string | null = null;
-let filterByEvalName: string | null = null;
-
-/**
- * Check the first argument:
- * - If it is "category", the next argument should be the category name.
- * - Otherwise, assume it is a specific evaluation (task) name.
- */
-if (args.length > 0) {
-  if (args[0].toLowerCase() === "category") {
-    filterByCategory = args[1];
-    if (!filterByCategory) {
-      console.error("Error: Category name not specified.");
-      process.exit(1);
-    }
-    // Validate that the category is one of the known ones.
-    try {
-      EvalCategorySchema.parse(filterByCategory);
-    } catch {
-      console.error(
-        `Error: Invalid category "${filterByCategory}". Valid categories are: ${DEFAULT_EVAL_CATEGORIES.join(", ")}`,
-      );
-      process.exit(1);
-    }
-  } else {
-    // Otherwise, treat it as a filter by evaluation name.
-    filterByEvalName = args[0];
-  }
-}
-
-/**
- * The `tasksConfig` defines all tasks from the config file. Each task has a name and categories.
- * We create a mapping `tasksByName` from task name to its categories for quick lookup.
- */
-type TaskConfig = { name: string; categories: string[] };
-const tasksConfig = config.tasks as TaskConfig[];
-
-const tasksByName = tasksConfig.reduce<
-  Record<string, { categories: string[] }>
->((acc, task) => {
-  acc[task.name] = { categories: task.categories };
-  return acc;
-}, {});
-
-/**
- * If filtering by a specific eval name (task), ensure that this task actually exists.
- */
-if (filterByEvalName && !tasksByName[filterByEvalName]) {
-  console.error(`Error: Evaluation "${filterByEvalName}" does not exist.`);
-  process.exit(1);
-}
-
-/**
- * Determine which models to run the evaluations against.
- *
- * DEFAULT_EVAL_MODELS: The default set of models used for most categories.
- * EXPERIMENTAL_EVAL_MODELS: Additional models included if the category is "experimental".
- */
-const DEFAULT_EVAL_MODELS = process.env.EVAL_MODELS
-  ? process.env.EVAL_MODELS.split(",")
-  : ["gpt-4o", "claude-3-5-sonnet-latest"];
-
-const EXPERIMENTAL_EVAL_MODELS = process.env.EXPERIMENTAL_EVAL_MODELS
-  ? process.env.EXPERIMENTAL_EVAL_MODELS.split(",")
-  : ["o1-mini", "o1-preview"];
-
-/**
- * getModelList:
- * Returns a list of models to be used for the given category.
- * If category is "experimental", it merges DEFAULT_EVAL_MODELS and EXPERIMENTAL_EVAL_MODELS.
- * Otherwise, returns DEFAULT_EVAL_MODELS.
- */
-const getModelList = (category: string | null): string[] => {
-  if (category === "experimental") {
-    // Remove duplicates by creating a Set and converting back to array.
-    return Array.from(
-      new Set([...DEFAULT_EVAL_MODELS, ...EXPERIMENTAL_EVAL_MODELS]),
-    );
-  }
-  return DEFAULT_EVAL_MODELS;
-};
-
-/**
- * MODELS: Final list of models that will be tested. We validate each model against `AvailableModelSchema`
- * to ensure they are supported.
- */
-const MODELS: AvailableModel[] = getModelList(filterByCategory).map((model) => {
-  if (!AvailableModelSchema.safeParse(model).success) {
-    throw new Error(`Model ${model} is not a supported model`);
-  }
-  return model as AvailableModel;
-});
 
 /**
  * generateSummary:
@@ -367,7 +230,7 @@ const generateFilteredTestcases = (): Testcase[] => {
             console.log(`❌ ${input.name}: Failed`);
           }
           return result;
-        } catch (error: any) {
+        } catch (error) {
           // Log any errors that occur during task execution
           console.error(`❌ ${input.name}: Error - ${error}`);
           logger.error({
